NIFI-16125 Split initial validation sweep into a dedicated ParallelTriggerValidationTask
Addresses review feedback from exceptionfactory: TriggerValidationTask
previously branched on whether an ExecutorService was supplied to decide
between the one-time parallel startup sweep and the periodic serial
re-validation. ParallelTriggerValidationTask now owns the startup sweep
exclusively; TriggerValidationTask is restored to its original serial-only
behavior for the periodic re-validation task.
